Split up model files and rename service to view model
