+++ src/xcode/ENA/ENA/Source/Scenes/ContactDiary/Model/ContactPersonEncounter.swift
@@ -1,0 +1,16 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import Combine
+
+struct ContactPersonEncounter {
+
+	// MARK: - Internal
+
+	let id: Int
+	let date: String
+	let contactPersonId: Int
+	
+}
