+++ src/xcode/ENA/ENA/Source/Scenes/ContactDiary/Model/DiaryEntryType.swift
@@ -1,0 +1,15 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import Combine
+
+enum DiaryEntryType {
+
+	// MARK: - Internal
+
+	case location
+	case contactPerson
+
+}
