+++ src/xcode/ENA/ENA/Source/Scenes/ContactDiary/Model/LocationVisit.swift
@@ -1,0 +1,16 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import Combine
+
+struct LocationVisit {
+
+	// MARK: - Internal
+
+	let id: Int
+	let date: String
+	let locationId: Int
+
+}
