+++ src/xcode/ENA/ENA/Source/Scenes/ContactDiary/Store/DiaryStoringProviding.swift
@@ -1,0 +1,49 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import Combine
+
+typealias DiaryStoringProviding = DiaryStoring & DiaryProviding
+
+protocol DiaryStoring {
+
+	typealias DiaryStoringResult = Result<Int, SQLiteErrorCode>
+	typealias DiaryStoringVoidResult = Result<Void, SQLiteErrorCode>
+
+	@discardableResult
+	func addContactPerson(name: String) -> DiaryStoringResult
+	@discardableResult
+	func addLocation(name: String) -> DiaryStoringResult
+	@discardableResult
+	func addContactPersonEncounter(contactPersonId: Int, date: String) -> DiaryStoringResult
+	@discardableResult
+	func addLocationVisit(locationId: Int, date: String) -> DiaryStoringResult
+
+	@discardableResult
+	func updateContactPerson(id: Int, name: String) -> DiaryStoringVoidResult
+	@discardableResult
+	func updateLocation(id: Int, name: String) -> DiaryStoringVoidResult
+
+	@discardableResult
+	func removeContactPerson(id: Int) -> DiaryStoringVoidResult
+	@discardableResult
+	func removeLocation(id: Int) -> DiaryStoringVoidResult
+	@discardableResult
+	func removeContactPersonEncounter(id: Int) -> DiaryStoringVoidResult
+	@discardableResult
+	func removeLocationVisit(id: Int) -> DiaryStoringVoidResult
+	@discardableResult
+	func removeAllLocations() -> DiaryStoringVoidResult
+	@discardableResult
+	func removeAllContactPersons() -> DiaryStoringVoidResult
+	@discardableResult
+	func cleanup() -> DiaryStoringVoidResult
+}
+
+protocol DiaryProviding {
+
+	var diaryDaysPublisher: CurrentValueSubject<[DiaryDay], Never> { get }
+	
+}
